requirements.txt
streamlit>=1.29.0
Cope>=2.2.2
streamlit_extras
streamlit_code_editor
sympy
ezregex
numpy
